delete rustref.go; add readme to sd
