+++ sd/README.md
@@ -1,0 +1,2 @@
+## `sd` package
+
